// src/shard/enums.ts
/************************************
 * 색상 유형
 ************************************/
export enum ColorVariants {
  contrast = "contrast",
  destructive = "destructive",
  primary = "primary",
  secondary = "secondary",
}
/************************************
 * 크기 유형
 ************************************/
export enum SizeVariants {
  xxs = "2xs",
  xs = "xs",
  sm = "sm",
  md = "md",
  lg = "lg",
  xl = "xl",
  xxl = "2xl",
}